+++ packages/terra-collapsible-menu-view/src/CollapsibleMenuView.jsx
@@ -77,8 +77,15 @@ class CollapsibleMenuView extends React.Component {
 
     for (let i = 0; i < this.props.children.length; i += 1) {
       const child = this.container.children[i];
-      calcWidth += child.getBoundingClientRect().width;
+      const childWidth = child.getBoundingClientRect().width;
+      calcWidth += childWidth;
+
       if (calcWidth > availableWidth) {
+        // If last child fits in the available space, leave it face up
+        if (i === this.props.children.length - 1 && calcWidth <= width) {
+          break;
+        }
+
         hiddenStartIndex = i;
         menuHidden = false;
         break;
@@ -91,6 +98,7 @@ class CollapsibleMenuView extends React.Component {
   render() {
     const { children, boundingRef, menuWidth, ...customProps } = this.props;
     const visibleChildren = React.Children.toArray(children);
+
     let hiddenChildren = null;
     if (this.state.hiddenStartIndex >= 0) {
       hiddenChildren = visibleChildren.splice(this.state.hiddenStartIndex);