+++ packages/terra-i18n-plugin/CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog
 
 Unreleased
 ----------
+* Allow new config options: translationsDirectoryRouters as recursive search transit and translationsDirectories as absolute search path
 
 1.8.0 - (October 24, 2017)
 ------------------